# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Firefly Airline",
        "title": "Firefly Airline API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/airports": {
            "get": {
                "security": [
                    {
                        "AgencyAPIKey": []
                    }
                ],
                "description": "Returns airports, optionally filtered by city.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airports"
                ],
                "summary": "List airports",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by city ID",
                        "name": "city_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/firefly-airline_internal_api.APIResponse-array_internal_domains_airport_AirportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/cities": {
            "get": {
                "security": [
                    {
                        "AgencyAPIKey": []
                    }
                ],
                "description": "Returns cities, optionally filtered by province.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "List cities",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by province ID",
                        "name": "province_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/firefly-airline_internal_api.APIResponse-array_internal_domains_city_CityResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/countries": {
            "get": {
                "security": [
                    {
                        "AgencyAPIKey": []
                    }
                ],
                "description": "Returns all countries.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Countries"
                ],
                "summary": "List countries",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/firefly-airline_internal_api.APIResponse-array_internal_domains_country_CountryResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/internal/agencies": {
            "post": {
                "security": [
                    {
                        "AirlineAPIKey": []
                    }
                ],
                "description": "Creates a new travel agency and generates an API key for it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agencies"
                ],
                "summary": "Create a new agency",
                "parameters": [
                    {
                        "description": "Agency information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domains_agency.CreateAgencyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/firefly-airline_internal_api.APIResponse-internal_domains_agency_CreateAgencyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/internal/airplanes": {
            "get": {
                "security": [
                    {
                        "AirlineAPIKey": []
                    }
                ],
                "description": "Returns all airplanes.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airplanes"
                ],
                "summary": "List airplanes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_domains_airplane.AirplaneResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AirlineAPIKey": []
                    }
                ],
                "description": "Creates a new airplane.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airplanes"
                ],
                "summary": "Create an airplane",
                "parameters": [
                    {
                        "description": "Airplane information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domains_airplane.CreateAirplaneRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_domains_airplane.AirplaneResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/internal/airplanes/{id}": {
            "get": {
                "security": [
                    {
                        "AirlineAPIKey": []
                    }
                ],
                "description": "Gets an airplane by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airplanes"
                ],
                "summary": "Get an airplane",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Airplane ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domains_airplane.AirplaneResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AirlineAPIKey": []
                    }
                ],
                "description": "Deletes an airplane. Deletion fails if the airplane is referenced by a flight.",
                "tags": [
                    "Airplanes"
                ],
                "summary": "Delete an airplane",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Airplane ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "AirlineAPIKey": []
                    }
                ],
                "description": "Updates an airplane.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airplanes"
                ],
                "summary": "Update an airplane",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Airplane ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Airplane information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domains_airplane.UpdateAirplaneRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domains_airplane.AirplaneResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/internal/airports": {
            "get": {
                "security": [
                    {
                        "AirlineAPIKey": []
                    }
                ],
                "description": "Returns all airports, optionally filtered by city.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airports"
                ],
                "summary": "List airports",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by city ID",
                        "name": "city_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/firefly-airline_internal_api.APIResponse-array_internal_domains_airport_AirportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AirlineAPIKey": []
                    }
                ],
                "description": "Creates a new airport.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airports"
                ],
                "summary": "Create an airport",
                "parameters": [
                    {
                        "description": "Airport information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domains_airport.CreateAirportRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/firefly-airline_internal_api.APIResponse-internal_domains_airport_AirportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/internal/airports/{id}": {
            "get": {
                "security": [
                    {
                        "AirlineAPIKey": []
                    }
                ],
                "description": "Gets an airport by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airports"
                ],
                "summary": "Get an airport",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Airport ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/firefly-airline_internal_api.APIResponse-internal_domains_airport_AirportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "AirlineAPIKey": []
                    }
                ],
                "description": "Updates an airport.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airports"
                ],
                "summary": "Update an airport",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Airport ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Airport information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domains_airport.UpdateAirportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/firefly-airline_internal_api.APIResponse-internal_domains_airport_AirportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "firefly-airline_internal_api.APIResponse-array_internal_domains_airport_AirportResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_domains_airport.AirportResponse"
                    }
                }
            }
        },
        "firefly-airline_internal_api.APIResponse-array_internal_domains_city_CityResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_domains_city.CityResponse"
                    }
                }
            }
        },
        "firefly-airline_internal_api.APIResponse-array_internal_domains_country_CountryResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_domains_country.CountryResponse"
                    }
                }
            }
        },
...
        "firefly-airline_internal_api.APIResponse-internal_domains_agency_CreateAgencyResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/internal_domains_agency.CreateAgencyResponse"
                }
            }
        },
...
        "internal_domains_agency.CreateAgencyRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "walletBalance": {
                    "type": "integer"
                }
            }
        },
...
        "internal_domains_airplane.AirplaneResponse": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer",
                    "example": 180
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "type": {
                    "type": "string",
                    "example": "Airbus A320"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AgencyAPIKey": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "AirlineAPIKey": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}